+++ app/src/main/java/com/app/arcabyolimpo/domain/model/product/product.kt
@@ -1,15 +1,13 @@
 package com.app.arcabyolimpo.domain.model.product
 
-/**
- * Domain model para mostrar / listar productos en la app.
- * NO reemplaza ProductAdd, lo complementa.
- */
+
 data class Product(
-    val id: String,              // idProducto
-    val name: String,            // Nombre
-    val unitaryPrice: String,    // PrecioUnitario
-    val workshopName: String?,   // nombreTaller
-    val description: String?,    // Descripcion
-    val available: Boolean,      // Disponible == 1 -> true
-    val imageUrl: String?,       // imagen
+    val id: String,
+    val name: String,          
+    val unitaryPrice: String,    
+    val workshopName: String?,  
+    val description: String?,    
+    val available: Boolean,      
+    val imageUrl: String?,      
+    val category: String?,       
 )